+++ src/utils/chrome/__tests__/chrome-crawler.test.ts
@@ -1,40 +1,44 @@
-import { describe, expect, it } from "bun:test";
+// To add a extension to this test:
+//
+// 1. Uncomment this code inside `src/crawlers/chrome-crawler.ts`:
+//    https://github.com/wxt-dev/queue/blob/2c538514080e2c787d47cae9f70e9e948fdcbd8b/src/crawlers/chrome-crawler.ts#L25-L30
+// 2. Run server in dev mode and use the `chromeExtension` query to get the
+//    extension's details
+// 3. After that query, regardless of if it succeeded or not, a new HTML file
+//    will be added to `src/crawlers/__tests__/fixtures/chrome-web-store/.new`
+// 4. Move the HTML file up one folder so it's next to the other test fixtures
+// 5. You're done! The test is added, run `bun test`.
+//
+
+import { beforeEach, describe, expect, it, mock } from "bun:test";
 import { crawlExtension } from "../chrome-crawler";
+import { readdir } from "node:fs/promises";
+import { join } from "node:path";
 
-const githubBetterLineCountsId = "ocfdgncpifmegplaglcnglhioflaimkd";
+const fetchMock = mock<typeof fetch>();
+globalThis.fetch = fetchMock as any;
 
-describe("Chrome Web Store Crawler", () => {
-  it("should load and crawl an extension ID correctly", async () => {
-    const res = await crawlExtension(githubBetterLineCountsId, "en");
+describe("Chrome Web Store Crawler", async () => {
+  const fixturesDir = join(import.meta.dir, "fixtures/chrome-web-store");
+  const testFiles = (await readdir(fixturesDir))
+    .filter((file) => !file.startsWith("."))
+    .toSorted();
+  const getExtensionIdFromFile = (file: string): string =>
+    file.match(/.*-([a-z]+)\.html/)![1]!;
 
-    expect(res).toEqual({
-      iconUrl:
-        "https://lh3.googleusercontent.com/GcffNyCJaxT2G9dsQCJHhUEMlu_E0vEzph5cLPrQj7UHKat7QyCzGu69Dmp_DDUL8rY-bPMFJceQarS1wcqdwTalTg=s256",
-      id: githubBetterLineCountsId,
-      lastUpdated: expect.any(String),
-      longDescription: expect.stringContaining("Isn't it annoying when you"),
-      name: "GitHub: Better Line Counts",
-      rating: expect.any(Number),
-      reviewCount: expect.any(Number),
-      shortDescription: "Remove generated files from GitHub line counts",
-      storeUrl:
-        "https://chromewebstore.google.com/detail/github-better-line-counts/ocfdgncpifmegplaglcnglhioflaimkd",
-      version: expect.any(String),
-      weeklyActiveUsers: expect.any(Number),
-      screenshots: [
-        {
-          index: 0,
-          indexUrl:
-            "http://localhost:3000/api/rest/chrome-extensions/ocfdgncpifmegplaglcnglhioflaimkd/screenshots/0",
-          rawUrl: expect.any(String),
-        },
-        {
-          index: 1,
-          indexUrl:
-            "http://localhost:3000/api/rest/chrome-extensions/ocfdgncpifmegplaglcnglhioflaimkd/screenshots/1",
-          rawUrl: expect.any(String),
-        },
-      ],
-    });
+  beforeEach(() => {
+    fetchMock.mockReset();
   });
+
+  it.each(testFiles)(
+    "should extract extension details from %s",
+    async (file) => {
+      const id = getExtensionIdFromFile(file);
+      fetchMock.mockResolvedValueOnce(
+        new Response(Bun.file(join(fixturesDir, file))),
+      );
+      const res = await crawlExtension(id, "en", true);
+      expect(res).toMatchSnapshot();
+    },
+  );
 });