+++ src/utils/chrome/chrome-crawler.ts
@@ -5,6 +5,7 @@ import { buildScreenshotUrl } from "../urls";
 export async function crawlExtension(
   id: string,
   lang: string,
+  canGenerateTestFixture = false,
 ): Promise<Gql.ChromeExtension | undefined> {
   consola.info("Crawling " + id);
   const url = `https://chromewebstore.google.com/detail/${id}?hl=${lang}`;
@@ -19,125 +20,129 @@ export async function crawlExtension(
   if (res.status !== 200) return;
 
   const html = await res.text();
-  const { document } = parseHTML(html);
+  if (!canGenerateTestFixture) {
+    // Uncomment to debug HTML or generate new test fixture
+    // const date = new Date();
+    // const dateString = `${date.getFullYear()}-${(date.getMonth() + 1).toString().padStart(2, "0")}-${date.getDate().toString().padStart(2, "0")}`;
+    // Bun.write(
+    //   `src/crawlers/__tests__/fixtures/chrome-web-store/.new/${dateString}-${id}.html`,
+    //   html,
+    // );
+  }
 
-  // Uncomment to debug HTML
-  // Bun.write("chrome.html", document.documentElement.outerHTML);
+  const parsed = parseHTML(html);
+  const document = parsed.document as HTMLElement;
 
-  // Basic metadata
-  const name = metaContent(document, "property=og:title")?.replace(
-    / - Chrome Web Store$/,
-    "",
-  );
-  const storeUrl = metaContent(document, "property=og:url");
-  const iconUrl = metaContent(document, "property=og:image")?.replace(
-    /=.+?$/,
-    "=s256",
-  );
-  const shortDescription = metaContent(document, "property=og:description");
+  if (document.querySelector("section") == null) {
+    // Extension is unavailable, ID no longer exists
+    return undefined;
+  }
 
-  // Grab the main sections that contain content
-  const sections = (document as HTMLElement).querySelectorAll(
-    "main > * > section",
+  const name = tryExtract("name", validateNonEmptyString, [
+    () =>
+      metaContent(document, "property=og:title")?.replace(
+        / - Chrome Web Store$/,
+        "",
+      ),
+  ]);
+  const storeUrl = tryExtract("storeUrl", validateNonEmptyString, [
+    () => metaContent(document, "property=og:url"),
+  ]);
+  const iconUrl = tryExtract("iconUrl", validateNonEmptyString, [
+    () => metaContent(document, "property=og:image")?.replace(/=.+?$/, "=s256"),
+  ]);
+  const shortDescription = tryExtract(
+    "shortDescription",
+    validateNonEmptyString,
+    [() => metaContent(document, "property=og:description")],
   );
-  const header: HTMLElement = sections[0];
-  const description: HTMLElement = sections[2];
-  const details: HTMLElement = sections[3];
 
   // Header
 
-  // userRowCount.outerHTHML:
-  // <div>
-  //   <a>...</a>
-  //   <a>...</a>
-  //   73 users
-  // </div>
-  // Remove the anchors and extract "73" from the text content
-  const userCountRow = header.querySelector("div:first-child > div:last-child");
-  userCountRow
-    .querySelectorAll("a")
-    .forEach((anchor: HTMLAnchorElement) => anchor.remove());
-  const weeklyActiveUsers = (userCountRow.textContent as string)
-    // "XYZ+ users"
-    .replace(" users", "")
-    .replace(",", "")
-    .replace("+", "")
-    .trim();
+  const weeklyActiveUsers = tryExtract("weeklyActiveUsers", validateInt, [
+    () => {
+      const userCountRow = document.querySelector(
+        "main > * > section:first-child > section > div > div:last-child",
+      ) as HTMLElement | null;
+      removeAnchorChildren(userCountRow);
+      return (
+        userCountRow?.textContent
+          // "W,XYZ+ users"
+          ?.replace(" users", "")
+          .replaceAll(",", "")
+          .replace("+", "")
+      );
+    },
+  ]);
 
-  // ratingRow.outerHTML:
-  // <span>
-  //   <span>
-  //     <span>5.0</span>
-  //     <svg ><path /></svg>
-  //     <span>(<a><p>2 ratings</p></a>)</span>
-  //   </span>
-  // </span>
-  const ratingRow = header.querySelector(
-    "div:first-child > div:nth-child(2) > span:last-child",
-  );
-  const rating =
-    ratingRow != null
-      ? extractNumber(
-          ratingRow.querySelector("span:first-child > span:first-child")
-            .textContent,
+  const rating = tryExtract("rating", validateFloat, [
+    () =>
+      document.querySelector(
+        "main > * > section:first-child > section > div > div:nth-child(2) > span > span > span",
+      )?.textContent,
+    () =>
+      document.querySelector("body").textContent.includes("No ratings")
+        ? 0
+        : undefined,
+  ]);
+
+  const reviewCount = tryExtract("reviewCount", validateInt, [
+    () =>
+      document
+        .querySelector(
+          "main > * > section:first-child > section > div > div:nth-child(2) > span > span > span:last-child > a",
         )
-      : 0;
-  const reviewCount =
-    ratingRow != null
-      ? extractNumber(ratingRow.querySelector("p").textContent)
-      : 0;
+        ?.textContent?.replace(" ratings", ""),
+    () =>
+      document.querySelector("body").textContent.includes("No ratings")
+        ? 0
+        : undefined,
+  ]);
 
-  // Details
+  const version = tryExtract("version", validateNonEmptyString, [
+    () => {
+      const listItems = [
+        ...document.querySelectorAll("main > * > section:nth-child(5) li"),
+      ];
+      const li = listItems.find((item) => item.textContent.includes("Version"));
+      return li?.querySelector(":scope > *:last-child")?.textContent;
+    },
+  ]);
 
-  const detailItems = details.querySelectorAll("li > div:last-child");
-  const version = detailItems[0].textContent.trim();
-  const lastUpdated = detailItems[1].textContent.trim();
+  const lastUpdated = tryExtract("lastUpdated", validateNonEmptyString, [
+    () => {
+      const listItems = [
+        ...document.querySelectorAll("main > * > section:nth-child(5) li"),
+      ];
+      const li = listItems.find((item) => item.textContent.includes("Updated"));
+      return li?.querySelector(":scope > *:last-child")?.textContent;
+    },
+  ]);
 
-  // Description
+  const longDescription = tryExtract(
+    "longDescription",
+    validateNonEmptyString,
+    [
+      () =>
+        document
+          .querySelector("main > * > section:nth-child(3) p:last-child")
+          ?.textContent?.replaceAll("\n\n", "\n"),
+    ],
+  );
 
-  const longDescription = description
-    .querySelector("p:last-child")
-    .textContent.replaceAll("\n\n", "\n");
-
-  // const longDescription = document
-  //   .querySelector("div[itemprop=description]")
-  //   ?.nextElementSibling?.textContent?.trim();
-  //
-  // const ratingDiv = document.querySelector(".rsw-stars");
-  // const rating = extractNumber(ratingDiv.title); // "Average rating: 4.78 stars"
-  // const reviewCount = extractNumber(ratingDiv.textContent); // "(1024)"
-
-  // <div
-  //   aria-label="Item media 1 screenshot"
-  //   data-media-url="https://lh3.googleusercontent.com/GUgh0ThX2FDPNvbaumYl4DqsUhsbYiCe-Hut9FoVEnkmTrXyA-sHbMk5jmZTj_t-dDP8rAmy6X6a6GNTCn9F8zo4VYU"
-  //   data-is-video="false"
-  //   data-slide-index="0"
-  // >
-  const screenshots = [...document.querySelectorAll("div[data-media-url]")]
-    .filter((div) => div.getAttribute("data-is-video") === "false")
-    .map<Gql.Screenshot>((div) => {
-      const index = Number(div.getAttribute("data-slide-index") || -1);
-      return {
-        index,
-        rawUrl: div.getAttribute("data-media-url") + "=s1280", // "s1280" gets the full resolution
-        indexUrl: buildScreenshotUrl("chrome-extensions", id, index),
-      };
-    });
-
-  if (name == null) return;
-  if (storeUrl == null) return;
-  if (iconUrl == null) return;
-  if (weeklyActiveUsers == null) return;
-  if (lastUpdated == null) return;
-  if (version == null) return;
-  if (shortDescription == null) return;
-  if (longDescription == null) return;
-  if (
-    screenshots.some(
-      (screenshot) => screenshot.index === -1 || !screenshot.rawUrl,
-    )
-  )
-    return;
+  const screenshots = tryExtract("screenshots", validateGqlScreenshots, [
+    () =>
+      [...document.querySelectorAll("div[data-media-url]")]
+        .filter((div) => div.getAttribute("data-is-video") === "false")
+        .map<Gql.Screenshot>((div) => {
+          const index = parseInt(div.getAttribute("data-slide-index"));
+          return {
+            index,
+            rawUrl: div.getAttribute("data-media-url") + "=s1280", // "s1280" gets the full resolution
+            indexUrl: buildScreenshotUrl("chrome-extensions", id, index),
+          };
+        }),
+  ]);
 
   const result: Gql.ChromeExtension = {
     id,
@@ -157,19 +162,79 @@ export async function crawlExtension(
   return result;
 }
 
-function metaContent(document: any, attrSelector: string): string | undefined {
+function metaContent(
+  document: HTMLElement,
+  attrSelector: string,
+): string | undefined {
   return document
     .querySelector(`meta[${attrSelector}]`)
     ?.getAttribute("content")
     .trim();
 }
 
-function extractNumber(text: string): number | undefined {
-  const res = /([0-9.,]+)/.exec(text)?.[1];
-  if (res == null) return;
+/** Try each of the different functions, collecting errors, and return the first value that can be parsed correctly. If no options succeed, return an error containing all the errors in it's cause. */
+function tryExtract<T>(
+  field: string,
+  validate: (value: any) => T,
+  extractors: (() => any)[],
+): T {
+  const errors: Error[] = [];
+  for (const extract of extractors) {
+    try {
+      const result = extract();
+      return validate(result);
+    } catch (error) {
+      errors.push(error as Error);
+    }
+  }
+  errors.forEach((err) => console.error(err));
+  throw new Error(`Could not extract "${field}" from HTML`, { cause: errors });
+}
 
-  const num = Number(res);
-  if (isNaN(num)) return;
+function validateInt(value: any): number {
+  const int = parseInt(value);
+  if (isNaN(int)) throw Error(`"${value}" was not an integer`);
+  return int;
+}
 
-  return num;
+function validateFloat(value: any): number {
+  const float = parseFloat(value);
+  if (isNaN(float)) throw Error(`"${value}" was not an float`);
+  return float;
+}
+
+function validateNonEmptyString(value: any): string {
+  if (typeof value !== "string" || value.trim().length === 0) {
+    throw new Error(`"${value}" was not a non-empty string`);
+  }
+  return value.trim();
+}
+
+function validateGqlScreenshots(value: any): Gql.Screenshot[] {
+  if (!Array.isArray(value)) {
+    throw new Error(`Screenshots must be an array`);
+  }
+  if (value.length === 0) {
+    throw new Error(`Screenshots array cannot be empty`);
+  }
+  return value.map((screenshot: any) => {
+    if (typeof screenshot !== "object" || screenshot === null) {
+      throw new Error(`Each screenshot must be an object`);
+    }
+    const index = validateInt(screenshot.index);
+    if (index < 0) throw Error("Screenshot index missing");
+    const rawUrl = validateNonEmptyString(screenshot.rawUrl);
+    const indexUrl = validateNonEmptyString(screenshot.indexUrl);
+    return {
+      index,
+      rawUrl,
+      indexUrl,
+    };
+  });
+}
+
+function removeAnchorChildren(element: HTMLElement | null | undefined): void {
+  element
+    ?.querySelectorAll("a")
+    .forEach((anchor: HTMLAnchorElement) => anchor.remove());
 }